Second Commit

+++ src/app/pages/user/user-page/user-page.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { User } from "../../../model/user";
-import { configUser } from "../../../config/user/configUser";
+import { configUser } from "../../../config/page/configUser";
 import { ActivatedRoute, Router } from "@angular/router";
 import { UsersService } from "../../../services/users/users.service";
 
@@ -29,7 +29,7 @@ export class UserPageComponent implements OnInit {
   btnClicked($event: any) {
     switch ($event.action) {
       case 'add':
-        this.router.navigate(['add/' + $event.item], {relativeTo: this.route}).then(r => ['']) ;
+        this.router.navigate(['add'], {relativeTo: this.route}).then(r => ['']) ;
         break;
       case 'edit':
         this.router.navigate(['detail/' + $event.item.id], {relativeTo: this.route}).then(r => ['']);
@@ -43,7 +43,9 @@ export class UserPageComponent implements OnInit {
 
   delete(user: User): void {
     this.users = this.users.filter(u => u !== user);
-    this.service.deleteUser(user.id).subscribe();
+    if (user.id !== undefined){
+      this.service.deleteUser(user.id).subscribe();
+    }
   }
 
   getUsers(): void {
